Massive update. Add lots request handlers. Precise cascade delete in configurations.

+++ SRC/Application/Players/Commands/CreatePlayer/CreatePlayerCommandValidator.cs
@@ -13,8 +13,8 @@ namespace BasketballLeague.Application.Players.Commands.CreatePlayer
             RuleFor(x => x.Surname)
                 .MinimumLength(3).WithMessage("Name minimum lenght is 3")
                 .NotEmpty().WithMessage("Surname field is required");
-            RuleFor(x => x.Position)
-                .InclusiveBetween((byte)1, (byte)5).WithMessage("Position must be be beetwen 1-5");
+            RuleFor(x => (int)x.Position)
+                .InclusiveBetween(1, 5).WithMessage("Position must be be beetwen 1-5");
         }
 
     }
