+++ SRC/Application/Players/Commands/UpdatePlayer/UpdatePlayerCommandValidator.cs
@@ -13,7 +13,7 @@ namespace BasketballLeague.Application.Players.Commands.UpdatePlayer
             RuleFor(x => x.Surname)
                 .MinimumLength(3).WithMessage("Name minimum lenght is 3")
                 .NotEmpty().WithMessage("Surname field is required");
-            RuleFor(x => x.Position)
+            RuleFor(x => (int)x.Position)
                 .InclusiveBetween(1, 5).WithMessage("Position must be be beetwen 1-5");
         }
     }
